Fixed sorghum panicle height (#290)
1. Fixing crash by empty scene with Vulkan Ray Tracer.
2. Refactored code, updated DigitalAgricultureProject
3. Fixed panicle height.

+++ EvoEngine_SDK/include/Rendering/Geometry/Mesh.hpp
@@ -235,14 +235,14 @@ class Mesh final : public IAsset, public IGeometry {
    *
    * @return The number of vertices.
    */
-  [[nodiscard]] size_t GetVerticesAmount() const;
+  [[nodiscard]] uint32_t GetVerticesAmount() const;
 
   /**
    * @brief Retrieves the number of triangles in the mesh.
    *
    * @return The number of triangles.
    */
-  [[nodiscard]] size_t GetTriangleAmount() const;
+  [[nodiscard]] uint32_t GetTriangleAmount() const;
 
   /**
    * @brief Recalculates the normals for the mesh.
